Adjust the limit for system IDs
Currently the system-id limit is set to 36 bytes which corresponds to
the length of a serialized UUID. However, the system-id limit in the
vswitch schema isn't limited by that and allows for arbitrary
identifiers that distinguish different OVS instances:

https://github.com/openvswitch/ovs/blob/ff55e8f3858ae153ef1996df3d52938f3df6d950/vswitchd/vswitch.ovsschema#L27-L29
https://docs.openvswitch.org/en/latest/ref/ovs-ctl.8/#options

In some cases deployments rely on FQDNs for system IDs which causes
labels set by exporters to be set to "unknown".

In order to avoid a completely arbitrary length for this exporter, a
limit of 253 bytes (max domain name length per RFC 1035) is introduced
by this change instead of a hard-coded length of 36.