Handle 2 whitespaces in dpctl/show output
It appears to be that in the current versions of OVS there are 2
whitespaces present on lines following the datapath line.

Not handling this case results in having errors like

'dpctl/show' command return for vswitchd-service failed output analysis
source="ovs_exporter.go:627"

and the lack of OVS datapath metrics present at the metrics endpoint.